Made ldtk::color module for color serialization/deserialization (might change this to utils later)
